requirements.txt
requests~=2.31.0
lxml~=5.1.0
beautifulsoup4~=4.12.3
selenium~=4.18.1
aiogram~=3.4.1
environs~=11.0.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.29
pandas~=2.2.1
PyYAML~=6.0.1